refactor(structure): Move exercises around different katas
This change moves the parsing expressions to kata2 and all the tests
for numbers to kata3. That's going to make the process a bit more
gradual.

Also, since the exercises can move from one kata to another, from now
one, all the tests will look for implementations only in a single
module called "api". You can save the following snippet within a file
called "api.ts" in your project if you were using the old structure
and don't want to change it for now:

  // api.ts
  export { BaseParser, ParsingError } from "./kata1";
  export { PrimitiveParser } from "./kata2";

And to confirm it still works, as always, run the command:

  $ npm test